Refactor `Coordinator`
